updated some sdks

+++ contactscommon/src/com/android/contacts/common/ContactsUtils.java
@@ -20,22 +20,19 @@ import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
 import android.net.Uri;
-import android.os.Build;
 import android.provider.ContactsContract.CommonDataKinds.Im;
+import android.provider.ContactsContract.DisplayPhoto;
 import android.support.annotation.IntDef;
 import android.support.v4.os.BuildCompat;
-import android.provider.ContactsContract.DisplayPhoto;
-import android.telephony.PhoneNumberUtils;
 import android.text.TextUtils;
 import android.util.Pair;
 
+import com.android.contacts.common.compat.ContactsCompat;
+import com.android.contacts.common.compat.DirectoryCompat;
+import com.android.contacts.common.model.AccountTypeManager;
 import com.android.contacts.common.model.account.AccountWithDataSet;
 import com.android.contacts.common.model.dataitem.ImDataItem;
 import com.android.contacts.common.testing.NeededForTesting;
-import com.android.contacts.common.compat.ContactsCompat;
-import com.android.contacts.common.compat.DirectoryCompat;
-import com.android.contacts.common.compat.SdkSelectionUtils;
-import com.android.contacts.common.model.AccountTypeManager;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -101,8 +98,8 @@ public class ContactsUtils {
     }
 
 
-    public static final long USER_TYPE_CURRENT = 0;
-    public static final long USER_TYPE_WORK = 1;
+    public static final int USER_TYPE_CURRENT = 0;
+    public static final int USER_TYPE_WORK = 1;
 
     /**
      * UserType indicates the user type of the contact. If the contact is from Work User (Work
